+++ src/canisters/frontend/ic-asset/src/asset/content.rs
@@ -1,4 +1,5 @@
 use crate::asset::content_encoder::ContentEncoder;
+use brotli::CompressorWriter;
 use dfx_core::error::fs::FsError;
 use flate2::write::GzEncoder;
 use flate2::Compression;
@@ -28,6 +29,7 @@ impl Content {
     pub fn encode(&self, encoder: &ContentEncoder) -> Result<Content, std::io::Error> {
         match encoder {
             ContentEncoder::Gzip => self.to_gzip(),
+            ContentEncoder::Brotli => self.to_brotli(),
             ContentEncoder::Identity => Ok(self.clone()),
         }
     }
@@ -38,6 +40,19 @@ impl Content {
         let data = e.finish()?;
         Ok(Content {
             data,
+            media_type: self.media_type.clone(),
+        })
+    }
+
+    pub fn to_brotli(&self) -> Result<Content, std::io::Error> {
+        let mut compressed_data = Vec::new();
+        {
+            let mut compressor = CompressorWriter::new(&mut compressed_data, 4096, 11, 22);
+            compressor.write_all(&self.data)?;
+            compressor.flush()?;
+        }
+        Ok(Content {
+            data: compressed_data,
             media_type: self.media_type.clone(),
         })
     }