+++ src/canisters/frontend/ic-asset/src/asset/content_encoder.rs
@@ -4,6 +4,8 @@ use serde::{Deserialize, Serialize};
 #[serde(rename_all = "lowercase")]
 pub enum ContentEncoder {
     Gzip,
+    #[serde(alias = "br")]
+    Brotli,
     Identity,
 }
 
@@ -11,6 +13,7 @@ impl std::fmt::Display for ContentEncoder {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match &self {
             ContentEncoder::Gzip => f.write_str("gzip"),
+            ContentEncoder::Brotli => f.write_str("br"),
             ContentEncoder::Identity => f.write_str("identity"),
         }
     }